// schemaregistry/rules/encryption/tink/proto/aes_gcm_pb.ts
// Copyright 2017 Google Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//
////////////////////////////////////////////////////////////////////////////////

// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
// @generated from file rules/encryption/tink/proto/aes_gcm.proto (package google.crypto.tink, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file rules/encryption/tink/proto/aes_gcm.proto.
 */
export const file_rules_encryption_tink_proto_aes_gcm: GenFile = /*@__PURE__*/
  fileDesc("CilydWxlcy9lbmNyeXB0aW9uL3RpbmsvcHJvdG8vYWVzX2djbS5wcm90bxISZ29vZ2xlLmNyeXB0by50aW5rIjQKD0Flc0djbUtleUZvcm1hdBIQCghrZXlfc2l6ZRgCIAEoDRIPCgd2ZXJzaW9uGAMgASgNIi8KCUFlc0djbUtleRIPCgd2ZXJzaW9uGAEgASgNEhEKCWtleV92YWx1ZRgDIAEoDEJYChxjb20uZ29vZ2xlLmNyeXB0by50aW5rLnByb3RvUAFaLWdpdGh1Yi5jb20vZ29vZ2xlL3RpbmsvcHJvdG8vYWVzX2djbV9nb19wcm90b6ICBlRJTktQQmIGcHJvdG8z");

/**
 * @generated from message google.crypto.tink.AesGcmKeyFormat
 */
export type AesGcmKeyFormat = Message<"google.crypto.tink.AesGcmKeyFormat"> & {
  /**
   * @generated from field: uint32 key_size = 2;
   */
  keySize: number;

  /**
   * @generated from field: uint32 version = 3;
   */
  version: number;
};

/**
 * Describes the message google.crypto.tink.AesGcmKeyFormat.
 * Use `create(AesGcmKeyFormatSchema)` to create a new message.
 */
export const AesGcmKeyFormatSchema: GenMessage<AesGcmKeyFormat> = /*@__PURE__*/
  messageDesc(file_rules_encryption_tink_proto_aes_gcm, 0);

/**
 * @generated from message google.crypto.tink.AesGcmKey
 */
export type AesGcmKey = Message<"google.crypto.tink.AesGcmKey"> & {
  /**
   * @generated from field: uint32 version = 1;
   */
  version: number;

  /**
   * @generated from field: bytes key_value = 3;
   */
  keyValue: Uint8Array;
};

/**
 * Describes the message google.crypto.tink.AesGcmKey.
 * Use `create(AesGcmKeySchema)` to create a new message.
 */
export const AesGcmKeySchema: GenMessage<AesGcmKey> = /*@__PURE__*/
  messageDesc(file_rules_encryption_tink_proto_aes_gcm, 1);