// schemaregistry/rules/encryption/tink/proto/aes_siv_pb.ts
// Copyright 2017 Google Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//
////////////////////////////////////////////////////////////////////////////////

// @generated by protoc-gen-es v2.0.0 with parameter "target=ts"
// @generated from file rules/encryption/tink/proto/aes_siv.proto (package google.crypto.tink, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file rules/encryption/tink/proto/aes_siv.proto.
 */
export const file_rules_encryption_tink_proto_aes_siv: GenFile = /*@__PURE__*/
  fileDesc("CilydWxlcy9lbmNyeXB0aW9uL3RpbmsvcHJvdG8vYWVzX3Npdi5wcm90bxISZ29vZ2xlLmNyeXB0by50aW5rIjQKD0Flc1NpdktleUZvcm1hdBIQCghrZXlfc2l6ZRgBIAEoDRIPCgd2ZXJzaW9uGAIgASgNIi8KCUFlc1NpdktleRIPCgd2ZXJzaW9uGAEgASgNEhEKCWtleV92YWx1ZRgCIAEoDEJPChxjb20uZ29vZ2xlLmNyeXB0by50aW5rLnByb3RvUAFaLWdpdGh1Yi5jb20vZ29vZ2xlL3RpbmsvcHJvdG8vYWVzX3Npdl9nb19wcm90b2IGcHJvdG8z");

/**
 * @generated from message google.crypto.tink.AesSivKeyFormat
 */
export type AesSivKeyFormat = Message<"google.crypto.tink.AesSivKeyFormat"> & {
  /**
   * Only valid value is: 64.
   *
   * @generated from field: uint32 key_size = 1;
   */
  keySize: number;

  /**
   * @generated from field: uint32 version = 2;
   */
  version: number;
};

/**
 * Describes the message google.crypto.tink.AesSivKeyFormat.
 * Use `create(AesSivKeyFormatSchema)` to create a new message.
 */
export const AesSivKeyFormatSchema: GenMessage<AesSivKeyFormat> = /*@__PURE__*/
  messageDesc(file_rules_encryption_tink_proto_aes_siv, 0);

/**
 * key_type: type.googleapis.com/google.crypto.tink.AesSivKey
 *
 * @generated from message google.crypto.tink.AesSivKey
 */
export type AesSivKey = Message<"google.crypto.tink.AesSivKey"> & {
  /**
   * @generated from field: uint32 version = 1;
   */
  version: number;

  /**
   * First half is AES-CTR key, second is AES-SIV.
   *
   * @generated from field: bytes key_value = 2;
   */
  keyValue: Uint8Array;
};

/**
 * Describes the message google.crypto.tink.AesSivKey.
 * Use `create(AesSivKeySchema)` to create a new message.
 */
export const AesSivKeySchema: GenMessage<AesSivKey> = /*@__PURE__*/
  messageDesc(file_rules_encryption_tink_proto_aes_siv, 1);